+++ plugins/pa11y/src/index.ts
@@ -1,0 +1,5 @@
+import Pa11y from './tasks/pa11y'
+
+export const tasks = [
+  Pa11y
+]