+++ plugins/heroku/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @dotcom-tool-kit/npm bumped from ^2.0.1 to ^2.0.2
+    * @dotcom-tool-kit/types bumped from ^2.0.0 to ^2.1.0
+    * @dotcom-tool-kit/vault bumped from ^2.0.0 to ^2.0.1
+
 ### [2.0.1](https://github.com/Financial-Times/dotcom-tool-kit/compare/heroku-v2.0.0...heroku-v2.0.1) (2022-05-03)