+++ plugins/nodemon/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @dotcom-tool-kit/types bumped from ^2.0.0 to ^2.1.0
+    * @dotcom-tool-kit/vault bumped from ^2.0.0 to ^2.0.1
+
 ## [2.0.0](https://github.com/Financial-Times/dotcom-tool-kit/compare/nodemon-v1.9.0...nodemon-v2.0.0) (2022-04-19)